[Refactor] Create New AbstractWebHookNotifier
The implementation for a ms-teams notifier is nearly identical to the
implementation of the slack notifier so I decided to create a new base
class to handle Webhook Notifiers
