+++ hooks/notification/Notifiers/SlackNotifier.ts
@@ -3,30 +3,17 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { NotifierType } from "../NotifierType"
-import { AbstractNotifier } from "./AbstractNotifier"
+import { AbstractWebHookNotifier } from "./AbstractWebHookNotifier"
 import { Finding } from "../model/Finding"
 import axios from 'axios';
 import { NotificationChannel } from "../model/NotificationChannel";
 import { Scan } from "../model/Scan";
 
-export class SlackNotifier extends AbstractNotifier {
+export class SlackNotifier extends AbstractWebHookNotifier {
 
   protected type: NotifierType = NotifierType.SLACK
 
   constructor(channel: NotificationChannel, scan: Scan, findings: Finding[], args: Object) {
     super(channel, scan, findings, args);
-  }
-
-  public async sendMessage(): Promise<void> {
-    await this.sendPostRequest(this.renderMessage());
-  }
-
-  protected async sendPostRequest(message: string) {
-    try {
-      await axios.post(this.channel.endPoint, message)
-    } catch (e) {
-      console.log(`There was an Error sending the Message for the Slack Notifier "${this.channel.name}"`);
-      console.log(e);
-    }
   }
 }
